+++ database/migrations/2024_09_24_043324_create_products_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->integer('product_id');
+            $table->bigIncrements('product_id');
             $table->string('name');
             $table->string('display_name');
             $table->string('code');
